Minor
Mlie reported an issue in which the Fungi Race called on Endo-Gene Categories that did not exist. This should fix that problem. I also diversified the Backstories of the Faction a bit, added compatibility for more types of fungi, and double-checked to see if the Vanilla Factions Expanded Fungi Pawn Kind Def was rendered properly.